Metro Style Theme & Remembers Open File States Between Sessions
Added the MetroFramework library to the project to add a more modern look and feel.

Added support for auto saving the state of opened filed to the .nwscript folder inside the applications folder. If you close a window without saving the contents to disk it will save a copy to the .nwscript folder and open that version on startup.

Reworked quite a bit of the old code and moved some stuff into their own classes. This was needed to extend the application beyond the main form.

Changed the right info pane to a TabControl and made tabs for Functions and Constants. The removes the need for the Constants button.

The app is now smarter and remembers more of your choices between sessions.
A few examples are:
 - The Targeted Game
 - Tab Spacing
 - Selected Theme ( Light | Dark )
 - Open Files
 - Open File Cursor Position